ml-service/requirements.txt
flask
flask-cors
pymongo
pandas
numpy
scikit-learn
xgboost
joblib
python-dotenv